pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "python-dotenv>=1.1.0",
]
